replaced doub;e with decimal

+++ client/Lykke.Service.PayInternal.Client/Models/Transactions/UpdateTransactionRequest.cs
@@ -7,7 +7,7 @@ namespace Lykke.Service.PayInternal.Client.Models.Transactions
     {
         public string TransactionId { get; set; }
         public string WalletAddress { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public int Confirmations { get; set; }
         [CanBeNull] public string BlockId { get; set; }
         public DateTime? FirstSeen { get; set; }
